+++ index.html
@@ -123,7 +123,10 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
-    <script src="/js/index.js"></script>
+        <script type="module" src="./modulos/index.js"></script>
+        <script type="module" src="./modulos/module.js"></script>
+
+
 </body>
 
 </html>
