+++ pastEents.html
@@ -120,7 +120,10 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
-        <script src="/js/pastEvents.js"></script>
+       
+        <script type="module" src="./modulos/pastEvents.js"></script>
+        <script type="module" src="./modulos/module.js"></script>
+
 </body>
 
 </html>
