+++ stats.html
@@ -6,7 +6,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-        <link rel="icon" type="image/png" href="image/favicon.png">
+    <link rel="icon" type="image/png" href="image/favicon.png">
     <link rel="stylesheet" href="Styles/main.css">
     <title>Stats</title>
 </head>
@@ -77,62 +77,34 @@
                 <span class="visually-hidden">Next</span>
             </button>
         </div>
+
         <div class="container">
-            <table class="table border">
+            <h2>future Events</h2>
+            <table class="table">
                 <thead>
                     <tr>
-                        <th scope="col">Events Statistics</th>
+                        <th>Name</th>
+                        <th>Renueves</th>
+                        <th>Asitencia </th>
                     </tr>
                 </thead>
-                <tbody>
-                    <tr>
-                        <td>events whit hghest % of assitance</td>
-                        <td>events whit lowest% of assitance</td>
-                        <td>events whit lowest% of assitance</td>
-                    </tr>
-                    <tr>
-                        <td>hi </td>
-                        <td>hello</td>
-                        <td>hola</td>
-                    </tr>
-                    <thead>
-                        <tr>
-                            <th scope="col">Events with Highest Attendance</th>
-                        </tr>
-                    </thead>
-                </tbody>
-                <tbody>
-                    <tr>
-                        <td>categories</td>
-                        <td>reenues</td>
-                        <td>asstence</td>
-                    </tr>
-                    <tr>
-                        <td>loremm</td>
-                        <td>loremm</td>
-                        <td>loremm</td>
-                    </tr>
-                </tbody>
-                <thead>
-                    <tr>
-                        <th scope="col">Events with Highest Attendance</th>
-                    </tr>
-                </thead>
-                <tbody>
-                    <tr>
-                        <td>categories</td>
-                        <td>reenues</td>
-                        <td>asistence</td>
-                    </tr>
-                    <tr>
-                        <td>lore</td>
-                        <td>Elore</td>
-                        <td>asasdad</td>
-                    </tr>
+                <tbody id="tablaEventosFuturos">
                 </tbody>
             </table>
-       
-       
+
+            <h2>Past Events</h2>
+            <table class="table">
+                <thead>
+                    <tr>
+                        <th>Name</th>
+                        <th>Renueves</th>
+                        <th>Asitencia </th>
+                    </tr>
+                </thead>
+                <tbody id="tablaEventosPasados">
+                </tbody>
+            </table>
+        </div>
         </section>
 
 
@@ -141,12 +113,13 @@
     <footer>
         <div class="container-fluid row ">
             <div class=" container-fluid row col-6  justify-content-center ">
-                <div class="icoon  col-2  "><a href=""><img class="icoonImg" src="image/facebook.png" alt="facebook"></a>
+                <div class="icoon  col-2  "><a href=""><img class="icoonImg" src="image/facebook.png"
+                            alt="facebook"></a>
                 </div>
                 <div class="icoon  col-2 mr-2"><a href=""><img class="icoonImg" src="image/instagram.png"
                             alt="instagram"></a></div>
-                <div class="icoon  col-2 mr-2"><a href="image/whatsapp.png"><img class="icoonImg" src="image/whatsapp.png"
-                            alt="whatsapp"></a></div>
+                <div class="icoon  col-2 mr-2"><a href="image/whatsapp.png"><img class="icoonImg"
+                            src="image/whatsapp.png" alt="whatsapp"></a></div>
             </div>
             <div class="col-6 p-3 text-end ">
                 <p>Harold Andres Diaz</p>
@@ -161,7 +134,7 @@
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
         integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
         crossorigin="anonymous"></script>
-        <script src="/js/stats.js"></script>
+    <script src="/modulos/stats.js"></script>
 </body>
 
 </html>
